Fix a bunch of size mismatches.
Probably every single instance of span usage needs to be
reviewed: a lot of code appears to assume that e.g.
gsl::make_span(x, 5) refers to a 5-byte block, independent
of the type of x.

Additionally, it would maybe make sense to return spans
instead of sizes as that would clear any uncertainty about
whether the results count entries or bytes.  OTOH I'm not
sure how to do that cleanly in the presence of errors.

Anyway, this is not finished, so it remains on the branch
until we find time to finish it up.

ZEN-69